servidor-sensores/requirements.txt
blinker==1.7.0
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.0
paho-mqtt==2.1.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
werkzeug==3.0.2
zipp==3.20.2
pyserial==3.5
